dialects: (angle) move to separate dialect
