+++ inconspiquous/dialects/gate.py
@@ -1,10 +1,7 @@
 from __future__ import annotations
-import math
 from typing import ClassVar
 
 from xdsl.dialects.builtin import (
-    FloatAttr,
-    Float64Type,
     IndexType,
     IntegerAttr,
     AnyFloatConstr,
@@ -42,104 +39,9 @@ from xdsl.printer import Printer
 from xdsl.traits import ConstantLike, HasCanonicalizationPatternsTrait, Pure
 from xdsl.dialects.builtin import IntAttrConstraint
 
+from inconspiquous.dialects.angle import AngleAttr
 from inconspiquous.gates import GateAttr, SingleQubitGate, TwoQubitGate
 from inconspiquous.constraints import SizedAttributeConstraint
-
-
-@irdl_attr_definition
-class AngleAttr(ParametrizedAttribute):
-    """
-    Attribute that wraps around a float attr, implicitly keeping it in the range
-    [0, 2) and implicitly multiplying by pi
-    """
-
-    name = "gate.angle"
-    data: ParameterDef[FloatAttr[Float64Type]]
-
-    def __init__(self, f: float):
-        f_attr: FloatAttr[Float64Type] = FloatAttr(f % 2, 64)
-        super().__init__((f_attr,))
-
-    @property
-    def as_float_raw(self) -> float:
-        return self.data.value.data
-
-    @property
-    def as_float(self) -> float:
-        return self.as_float_raw * math.pi
-
-    @classmethod
-    def parse_parameters(cls, parser: AttrParser) -> tuple[FloatAttr[Float64Type]]:
-        with parser.in_angle_brackets():
-            is_negative = parser.parse_optional_punctuation("-") is not None
-            f = parser.parse_optional_number()
-            if f is None:
-                f = 1.0
-            if isinstance(f, int):
-                f = float(f)
-            if f == 0.0:
-                parser.parse_optional_keyword("pi")
-            else:
-                parser.parse_keyword("pi")
-            if is_negative:
-                f = -f
-            return (FloatAttr(f % 2, 64),)
-
-    def print_parameters(self, printer: Printer) -> None:
-        with printer.in_angle_brackets():
-            f = self.as_float_raw
-            if f == 0.0:
-                printer.print_string("0")
-            elif f == 1.0:
-                printer.print_string("pi")
-            else:
-                printer.print_string(f"{f}pi")
-
-    def __add__(self, other: AngleAttr) -> AngleAttr:
-        return AngleAttr(self.data.value.data + other.data.value.data)
-
-    def __sub__(self, other: AngleAttr) -> AngleAttr:
-        return AngleAttr(self.data.value.data - other.data.value.data)
-
-    def __neg__(self) -> AngleAttr:
-        return AngleAttr(-self.data.value.data)
-
-
-@irdl_attr_definition
-class AngleType(ParametrizedAttribute, TypeAttribute):
-    """
-    A type for runtime angle values.
-    """
-
-    name = "gate.angle_type"
-
-
-@irdl_op_definition
-class ConstantAngleOp(IRDLOperation):
-    """
-    Constant-like operation for producing angles
-    """
-
-    name = "gate.constant_angle"
-
-    angle = prop_def(AngleAttr)
-
-    out = result_def(AngleType)
-
-    assembly_format = "`` $angle attr-dict"
-
-    traits = traits_def(
-        ConstantLike(),
-        Pure(),
-    )
-
-    def __init__(self, angle: AngleAttr):
-        super().__init__(
-            properties={
-                "angle": angle,
-            },
-            result_types=(AngleType(),),
-        )
 
 
 @irdl_attr_definition
@@ -435,10 +337,8 @@ class XZOp(IRDLOperation):
 
 Gate = Dialect(
     "gate",
-    [ConstantAngleOp, ConstantGateOp, QuaternionGateOp, ComposeGateOp, XZSOp, XZOp],
+    [ConstantGateOp, QuaternionGateOp, ComposeGateOp, XZSOp, XZOp],
     [
-        AngleAttr,
-        AngleType,
         HadamardGate,
         XGate,
         YGate,
