+++ tests/filecheck/dialects/gate.mlir
@@ -1,41 +1,9 @@
 // RUN: QUOPT_ROUNDTRIP
 // RUN: QUOPT_GENERIC_ROUNDTRIP
 
-"test.op"() {angle = #gate.angle<0>} : () -> ()
-
-// CHECK: "test.op"() {angle = #gate.angle<0>} : () -> ()
-
-"test.op"() {angle = #gate.angle<pi>} : () -> ()
-
-// CHECK-NEXT: "test.op"() {angle = #gate.angle<pi>} : () -> ()
-
-"test.op"() {angle = #gate.angle<2pi>} : () -> ()
-
-// CHECK-NEXT: "test.op"() {angle = #gate.angle<0>} : () -> ()
-
-"test.op"() {angle = #gate.angle<0.5pi>} : () -> ()
-
-// CHECK-NEXT: "test.op"() {angle = #gate.angle<0.5pi>} : () -> ()
-
-"test.op"() {angle = #gate.angle<1.5pi>} : () -> ()
-
-// CHECK-NEXT: "test.op"() {angle = #gate.angle<1.5pi>} : () -> ()
-
-"test.op"() {angle = #gate.angle<2.5pi>} : () -> ()
-
-// CHECK-NEXT: "test.op"() {angle = #gate.angle<0.5pi>} : () -> ()
-
-"test.op"() {angle = #gate.angle<-0.5pi>} : () -> ()
-
-// CHECK-NEXT: "test.op"() {angle = #gate.angle<1.5pi>} : () -> ()
-
-"test.op"() {angle = #gate.angle<-pi>} : () -> ()
-
-// CHECK-NEXT: "test.op"() {angle = #gate.angle<pi>} : () -> ()
-
 "test.op"() {gate = #gate.h} : () -> ()
 
-// CHECK-NEXT: "test.op"() {gate = #gate.h} : () -> ()
+// CHECK: "test.op"() {gate = #gate.h} : () -> ()
 
 "test.op"() {gate = #gate.rz<pi>} : () -> ()
 
@@ -50,10 +18,6 @@
 
 // CHECK-NEXT: %{{.*}} = gate.constant #gate.cz
 // CHECK-GENERIC-NEXT: %{{.*}} = "gate.constant"() <{gate = #gate.cz}> : () -> !gate.type<2>
-
-// CHECK-NEXT: %{{.*}} = gate.constant_angle<pi>
-// CHECK-GENERIC-NEXT: %{{.*}} = "gate.constant_angle"() <{angle = #gate.angle<pi>}> : () -> !gate.angle_type
-%a = gate.constant_angle<pi>
 
 %zero = arith.constant 0 : i64
 %one = arith.constant 1 : i64
